Fix dispatchEmailNotifications()
Store customer is looked up, however it is not the user who runs this function but the application when paypal's IPN returns the success message.

Therefore it will always return a null value breaking the email. I have changed this to look up the orders email attribute to use this to send the email out.